requirements.txt
pyserial
sounddevice
numpy
Pillow
